+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 * 21.x.x
+  - add version string from git describe
   - add CCPi-Regularisation toolkit in unittests
   - show_geometry implemented to display AcquisitionGeometry objects, can be imported from utilities.display
+  - CentreOfRotationCorrector.image_sharpness implemented which finds the rotation axis offset by maximising sharpness of a single slice reconstruction
+  - CentreOfRotationCorrector.xcorr renamed to CentreOfRotationCorrector.xcorrelation
 
 * 21.1.0
   - Added TomoPhantom plugin to create 2D/3D + channel ImageData phantoms based on the TomoPhantom model library